Move dir fo global generated types into src/@types
