+++ src/domain/offer/tableTools/TableTools.tsx
@@ -7,7 +7,7 @@ import styles from './tableTools.module.scss';
 import Text from '../../../common/text/Text';
 import Chip from '../../../common/chip/Chip';
 import { APPLICATION_STATUS } from '../../../common/utils/consonants';
-import { ApplicationStatus } from '../../../../__generated__/globalTypes';
+import { ApplicationStatus } from '../../../@types/__generated__/globalTypes';
 
 export interface TableToolsProps {
   applicationDate: string;
